requirements.txt
autogen[gemini,ipython]
openai
pandas
numpy
scikit-learn
python-dotenv
flaml[automl]